+++ ios/Classes/X448DartPlugin.m
@@ -1,0 +1,8 @@
+#import <Flutter/Flutter.h>
+
+@interface X448DartPlugin : NSObject<FlutterPlugin>
+@end
+
+@implementation X448DartPlugin
++ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {}
+@end