+++ linux/include/x448_dart/x448_dart_plugin.h
@@ -1,0 +1,9 @@
+#ifndef FLUTTER_PLUGIN_X448_DART_PLUGIN_H_
+#define FLUTTER_PLUGIN_X448_DART_PLUGIN_H_
+#include <flutter_linux/flutter_linux.h>
+G_BEGIN_DECLS
+
+void X448DartPluginRegisterWithRegistrar(FlPluginRegistrar* registrar);
+
+G_END_DECLS
+#endif