+++ windows/x448_dart_plugin.cpp
@@ -1,0 +1,2 @@
+#include "x448_dart_plugin.h"
+void X448DartPluginRegisterWithRegistrar(FlutterDesktopPluginRegistrar* registrar) {}